refactoring scan.go
